Improve burner efficiency between stages
The burner was trying to grease up to 100 deg. in the boiling stage,
but there's no reason not to blast into it.  Also, for the flush
stage, the charts look funny with an artificial set temp.

Related, the burner was trying to grease down to 99 deg. in the cool
stage, but there's no reason not to drop past that into idle.

Fix this by adding in a separate `burner-override` signal the control
can set to clobber the effect of the PID comp before it goes into the
limiter.