bb_autoclave.dts:  failed experiment
This overlay fragment doesn't work properly.  Previous messing around
with overlays was all attempts to fix symptoms of my old BBB, which
had an old MMC flash & broke things.  Now that that's fixed, these
overlays can probably be removed for good, since the pins start up in
the mode I need now (though there are several seconds at the beginning
where the burner is on; this could be bad).