Add low-pass filter to temp inputs
Lots of noise esp. at lower temps, and we want more stable inputs for
PID controller.

Do this with a low-pass filter, filtering signals greater than 0.1 Hz.
This leads to a 0.5 deg. error when heating in simuation.

In sim, re-add randomness from commit e5032824.

The HAL low-pass filter is the one-liner discussed here.
https://www.embeddedrelated.com/showarticle/779.php