common.hal:  PID param tweaks
These tweaks pretty much take care of overshoot.  The large I
component was wiping out the D component when approaching w/in a
couple of degrees, and would turn on the burner partway, leading to an
unclean curve.

At 120 deg. set, oscillation btw. 119.7 and 119.9.  The I param may
need to be increased slightly.