Create composite temp signal from thermistor and pressure sensors
The thermistor input is unstable, and at high temps, where we want
good control, it wanders around causing fluctuations of 0.5 deg. or
more.  This might be fixed with PID settings or something, but the
nice, stable pressure signal makes a good replacement.

Since the pressure signal is only meaningful above 100 deg., use the
thermistor under 105, the pressure signal above 110, and blend them
while in between.

Fixes #5