common.hal:  tune real machine
These settings are fairly stable, but undershoot and can take a while
to get up to temp and stabilize.

The integral product pretty much settles at 0.5 at 121 deg.  It also
causes some instability and is slow to adjust itself (since the burner
is usually saturated until things stabilize) so I'm debating zeroing
this out and trying again at setting a burner.bias that varies
proportionally to the difference with ambient temp.