common.hal:  more real machine tweaks
Set a 0.5 burner bias so that things are close to zeroed out at 121
deg.  That takes the place of a portion of the I band, so the max I
can be reduced.  Finally, remove the `maxoutput` restriction so that
the I band will continue to adjust even when the burner is saturated.